format bluestyle

+++ src/OptimalBids.jl
@@ -31,7 +31,7 @@ Clears the market.
 function clear_market!(market::Market) end
 
 """
-    calculate_profit(market::Market) -> NamedTuple{(:cleared_volumes, :clearing_prices, :profit), Tuple{Vector{Int64}, Vector{Int64}, Vector{Int64}}}
+    calculate_profit(market::Market) -> NamedTuple{(:cleared_volumes, :clearing_prices, :profit), Tuple{Vector{Int}, Vector{Int}, Vector{Int}}}
 
 Retrieves strategic agent's cleared volumes and prices from the market and calculates per bid profit.
 """
